pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geneinsight"
version = "0.1.0"
description = "Topic modeling pipeline for gene sets with enrichment analysis"
readme = "README.md"
authors = [
    {name = "WL Chin", email = "wee.chin@health.wa.gov.au"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Bioinformatics",
]
dependencies = [
    "pandas>=2.0.0",
    "bertopic>=0.15.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.2",
    "openai>=1.1.0",
    "instructor>=0.3.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "gseapy>=1.1.0",
    "joblib>=1.3.0",
    "optuna>=3.3.0",
    "stringdb>=0.1.5",
    "pydantic>=2.4.0",
    "rich>=13.9.4",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "seaborn>=0.12.0",
    "colorcet>=3.0.0",
    "torch>=1.7.0",
    "ollama>=0.4.7",
    "umap-learn>=0.5.7",
    "plotly>=5.9.0",
    "sphinx>=4.3.0",
    "pillow>=9.0.0",
    "sphinx-wagtail-theme>=6.5.0",
    "sphinx-togglebutton>=0.3.2",
    "urllib3>=2.0.0",
    "numba>=0.60.0",
]

[project.urls]
"Homepage" = "https://github.com/wlchin/geneinsight"
"Bug Tracker" = "https://github.com/wlchin/geneinsight/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pytest-cov>=3.0.0",
]
report = [
    "umap-learn>=0.5.7",
    "plotly>=5.9.0",
    "sphinx>=4.3.0",
    "pillow>=9.0.0",
    "sphinx-wagtail-theme>=6.5.0",
    "sphinx-togglebutton>=0.3.2",
]
ollama = ["ollama>=0.4.7"]
all = ["ollama>=0.4.7", "umap-learn>=0.5.7", "plotly>=5.9.0", "sphinx>=4.3.0", "pillow>=9.0.0", "sphinx-wagtail-theme>=6.5.0", "sphinx-togglebutton>=0.3.2"]

[project.scripts]
geneinsight = "geneinsight.cli:main"
geneinsight-enrichment = "geneinsight.scripts.geneinsight_enrichment:main"
geneinsight-topic = "geneinsight.scripts.geneinsight_topic_modeling:main"
geneinsight-prompt = "geneinsight.scripts.geneinsight_prompt_gen:main"
geneinsight-api = "geneinsight.scripts.geneinsight_api_process:main"
geneinsight-hypergeometric = "geneinsight.scripts.geneinsight_hypergeometric:main"
geneinsight-filter = "geneinsight.scripts.geneinsight_filter:main"
geneinsight-meta = "geneinsight.scripts.geneinsight_meta:main"
geneinsight-counter = "geneinsight.scripts.geneinsight_counter:main"
geneinsight-report = "geneinsight.scripts.geneinsight_report_pipeline:main"
geneinsight-ontology = "geneinsight.ontology.workflow:main"
geneinsight-examples = "geneinsight.scripts.geneinsight_examples:main"

[tool.setuptools]
include-package-data = true
packages = {find = {exclude = ["tests"]}}

[tool.setuptools.package-data]
"geneinsight.report" = ["assets/*.png"]
"geneinsight.ontology.ontology_folders" = ["*.txt"]
"geneinsight.examples" = ["*.txt", "*.md"]
"geneinsight.embedding_model" = ["*.json", "*.txt", "tokenizer.json", "*.bin"]

[tool.setuptools.exclude-package-data]
"*" = ["tests/*"]
